fix(hal): use the file's (int,int) requestFrom overload; reconcile architecture.md
The overload change and the architecture.md edits were in a local WIP
commit dropped during the merge of main and the PR body still described
them. Re-apply: (int,int) casts as the rest of the file uses, the scan
mode bullets in architecture.md say what each probe puts on the bus,
the Arduino header notes the Linux twin's -2, and the CHANGELOG tells a
consumer of the old data-byte write what to use instead.

+++ src/hal/arduino/crumbs_i2c_arduino.cpp
@@ -268,7 +268,7 @@ extern "C" int crumbs_arduino_scan(void *user_ctx,
             // the target ACKs its address for a read and clocks a byte out.
             // Puts no data on the bus, but does consume one byte from a
             // register-addressed or stream-style device.
-            uint8_t got = wire->requestFrom(static_cast<uint8_t>(addr), static_cast<uint8_t>(1));
+            uint8_t got = wire->requestFrom((int)addr, (int)1); // (int,int) as elsewhere in this file
             while (wire->available())
                 (void)wire->read();
             present = (got > 0);
